Move showSaveAsDialog into notebook/menu.js
